mostly fix the access program
